requirements.txt
# Samvad XR — Dev B Dependencies
# ================================

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0

# Async HTTP client (replaces requests)
httpx>=0.25.0

# RAG knowledge base
chromadb>=0.4.0

# Environment config
python-dotenv>=1.0.0

# Multipart file upload to Sarvam
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
